Added a conversion to bool operator for IoBase. The log now uses the StaticString class. Debugger works with x64 (except backtrace & trace).

+++ src/system/kernel/linker/Elf32.cc
@@ -132,7 +132,7 @@ unsigned int Elf32::getLastAddress()
 {
 }
 
-const char *Elf32::lookupSymbol(unsigned int addr, uintptr_t *startAddr)
+const char *Elf32::lookupSymbol(uintptr_t addr, uintptr_t *startAddr)
 {
   if (!m_pSymbolTable || !m_pStringTable)
     return 0; // Just return null if we haven't got a symbol table.
